Use EDAlias in the SwitchProducer in raw2cluster
