+++ EventFilter/SiPixelRawToDigi/python/SiPixelRawToDigi_cfi.py
@@ -9,8 +9,16 @@ siPixelDigis = SwitchProducerCUDA(
 from Configuration.Eras.Modifier_phase1Pixel_cff import phase1Pixel
 phase1Pixel.toModify(siPixelDigis.cpu, UsePhase1=True)
 
-from EventFilter.SiPixelRawToDigi.siPixelDigisFromSoA_cfi import siPixelDigisFromSoA as _siPixelDigisFromSoA
 from Configuration.ProcessModifiers.gpu_cff import gpu
 gpu.toModify(siPixelDigis,
-    cuda = _siPixelDigisFromSoA.clone(digiSrc = "siPixelClustersPreSplitting")
+    cuda = cms.EDAlias(
+        siPixelDigiErrors = cms.VPSet(
+            cms.PSet(type = cms.string("DetIdedmEDCollection")),
+            cms.PSet(type = cms.string("SiPixelRawDataErroredmDetSetVector")),
+            cms.PSet(type = cms.string("PixelFEDChanneledmNewDetSetVector"))
+        ),
+        siPixelDigisClustersPreSplitting = cms.VPSet(
+            cms.PSet(type = cms.string("PixelDigiedmDetSetVector"))
+        )
+    )
 )
