+++ EventFilter/SiPixelRawToDigi/python/siPixelDigis_cff.py
@@ -3,6 +3,7 @@ import FWCore.ParameterSet.Config as cms
 from EventFilter.SiPixelRawToDigi.SiPixelRawToDigi_cfi import siPixelDigis
 from EventFilter.SiPixelRawToDigi.siPixelDigisSoAFromCUDA_cfi import siPixelDigisSoAFromCUDA as _siPixelDigisSoAFromCUDA
 from EventFilter.SiPixelRawToDigi.siPixelDigiErrorsSoAFromCUDA_cfi import siPixelDigiErrorsSoAFromCUDA as _siPixelDigiErrorsSoAFromCUDA
+from EventFilter.SiPixelRawToDigi.siPixelDigiErrorsFromSoA_cfi import siPixelDigiErrorsFromSoA as _siPixelDigiErrorsFromSoA
 
 siPixelDigisTask = cms.Task(siPixelDigis)
 
@@ -12,7 +13,12 @@ siPixelDigisSoA = _siPixelDigisSoAFromCUDA.clone(
 siPixelDigiErrorsSoA = _siPixelDigiErrorsSoAFromCUDA.clone(
     src = "siPixelClustersCUDAPreSplitting"
 )
-siPixelDigisTaskCUDA = cms.Task(siPixelDigisSoA, siPixelDigiErrorsSoA)
+siPixelDigiErrors = _siPixelDigiErrorsFromSoA.clone()
+siPixelDigisTaskCUDA = cms.Task(
+    siPixelDigisSoA,
+    siPixelDigiErrorsSoA,
+    siPixelDigiErrors
+)
 
 from Configuration.ProcessModifiers.gpu_cff import gpu
 _siPixelDigisTask_gpu = siPixelDigisTask.copy()
